# packages/backend/src/controllers/AuthController.ts
import { Controller, Post, Body, Route, Tags, Response, Request } from "tsoa";
import { SupabaseService } from "../services/SupabaseService";
import { loginSchema, registerSchema, createErrorResponse } from "../utils/validation";
import { z } from "zod";

interface LoginRequest {
  email: string;
  password: string;
}

interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

interface AuthResponse {
  user: {
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    emailVerified: boolean;
  };
  session: {
    access_token: string;
    refresh_token: string;
    expires_at: number;
  };
}

interface ErrorResponse {
  error: string;
  message: string;
}

@Route("auth")
@Tags("Authentication")
export class AuthController extends Controller {
  private supabaseService: SupabaseService;

  constructor() {
    super();
    this.supabaseService = new SupabaseService();
  }

  /**
   * User login
   */
  @Post("login")
  @Response<AuthResponse>(200, "Login successful")
  @Response<ErrorResponse>(400, "Invalid credentials")
  @Response<ErrorResponse>(401, "Authentication failed")
  public async login(@Body() requestBody: LoginRequest): Promise<AuthResponse | ErrorResponse> {
    try {
      // Validate input
      const validatedData = loginSchema.parse(requestBody);

      // Use Supabase Auth
      const { data, error } = await this.supabaseService.signIn(validatedData.email, validatedData.password);

      if (error) {
        this.setStatus(401);
        return {
          error: "Authentication Failed",
          message: error.message,
        };
      }

      if (!data.user || !data.session) {
        this.setStatus(401);
        return {
          error: "Authentication Failed",
          message: "Invalid response from authentication service",
        };
      }

      const response: AuthResponse = {
        user: {
          id: data.user.id,
          email: data.user.email!,
          firstName: data.user.user_metadata?.first_name || "",
          lastName: data.user.user_metadata?.last_name || "",
          emailVerified: data.user.email_confirmed_at ? true : false,
        },
        session: {
          access_token: data.session.access_token,
          refresh_token: data.session.refresh_token,
          expires_at: data.session.expires_at!,
        },
      };

      this.setStatus(200);
      return response;
    } catch (error) {
      if (error instanceof z.ZodError) {
        this.setStatus(400);
        return {
          error: "Validation Error",
          message: error.errors.map((e) => e.message).join(", "),
        };
      }

      this.setStatus(500);
      return {
        error: "Internal Server Error",
        message: "An unexpected error occurred",
      };
    }
  }

  /**
   * User registration
   */
  @Post("register")
  @Response<AuthResponse>(201, "Registration successful")
  @Response<ErrorResponse>(400, "Invalid input")
  @Response<ErrorResponse>(409, "User already exists")
  public async register(@Body() requestBody: RegisterRequest): Promise<AuthResponse | ErrorResponse> {
    try {
      // Validate input
      const validatedData = registerSchema.parse(requestBody);

      // Use Supabase Auth
      const { data, error } = await this.supabaseService.signUp(validatedData.email, validatedData.password, {
        firstName: validatedData.firstName || "",
        lastName: validatedData.lastName || "",
      });

      if (error) {
        this.setStatus(409);
        return {
          error: "Registration Failed",
          message: error.message,
        };
      }

      if (!data.user) {
        this.setStatus(500);
        return {
          error: "Registration Failed",
          message: "Failed to create user",
        };
      }

      const response: AuthResponse = {
        user: {
          id: data.user.id,
          email: data.user.email!,
          firstName: data.user.user_metadata?.first_name || validatedData.firstName,
          lastName: data.user.user_metadata?.last_name || validatedData.lastName,
          emailVerified: data.user.email_confirmed_at ? true : false,
        },
        session: data.session
          ? {
              access_token: data.session.access_token,
              refresh_token: data.session.refresh_token,
              expires_at: data.session.expires_at!,
            }
          : {
              access_token: "",
              refresh_token: "",
              expires_at: 0,
            },
      };

      this.setStatus(201);
      return response;
    } catch (error) {
      if (error instanceof z.ZodError) {
        this.setStatus(400);
        return {
          error: "Validation Error",
          message: error.errors.map((e) => e.message).join(", "),
        };
      }

      this.setStatus(500);
      return {
        error: "Internal Server Error",
        message: "An unexpected error occurred",
      };
    }
  }

  /**
   * Password reset request
   */
  @Post("forgot-password")
  @Response<{ message: string }>(200, "Password reset email sent")
  @Response<ErrorResponse>(400, "Invalid email")
  public async forgotPassword(@Body() requestBody: { email: string }): Promise<{ message: string } | ErrorResponse> {
    try {
      if (!requestBody.email || !/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(requestBody.email)) {
        this.setStatus(400);
        return {
          error: "Invalid Email",
          message: "Please provide a valid email address",
        };
      }

      // Use Supabase Auth
      const { data, error } = await this.supabaseService.resetPassword(requestBody.email);

      if (error) {
        this.setStatus(500);
        return {
          error: "Failed to Send Email",
          message: error.message,
        };
      }

      this.setStatus(200);
      return {
        message: "Password reset email sent successfully",
      };
    } catch (error) {
      this.setStatus(500);
      return {
        error: "Internal Server Error",
        message: "Failed to send password reset email",
      };
    }
  }

  /**
   * Password reset confirmation
   */
  @Post("reset-password")
  @Response<{ message: string }>(200, "Password reset successful")
  @Response<ErrorResponse>(400, "Invalid token or password")
  public async resetPassword(@Body() requestBody: { token: string; password: string }): Promise<{ message: string } | ErrorResponse> {
    try {
      if (!requestBody.token || !requestBody.password) {
        this.setStatus(400);
        return {
          error: "Missing Fields",
          message: "Token and password are required",
        };
      }

      // Validate password strength
      if (requestBody.password.length < 8) {
        this.setStatus(400);
        return {
          error: "Weak Password",
          message: "Password must be at least 8 characters long",
        };
      }

      // Use Supabase Auth to update password
      const { data, error } = await this.supabaseService.updatePassword(requestBody.token, requestBody.password);

      if (error) {
        this.setStatus(400);
        return {
          error: "Password Reset Failed",
          message: error.message,
        };
      }

      this.setStatus(200);
      return {
        message: "Password reset successfully",
      };
    } catch (error) {
      this.setStatus(500);
      return {
        error: "Internal Server Error",
        message: "Failed to reset password",
      };
    }
  }

  /**
   * Logout user
   */
  @Post("logout")
  @Response<{ message: string }>(200, "Logout successful")
  public async logout(): Promise<{ message: string }> {
    // Use Supabase Auth
    const { error } = await this.supabaseService.signOut();

    if (error) {
      console.error("Logout error:", error);
    }

    this.setStatus(200);
    return {
      message: "Logged out successfully",
    };
  }

  /**
   * Update user profile
   */
  @Post("profile")
  @Response<{ message: string }>(200, "Profile updated successfully")
  @Response<ErrorResponse>(400, "Invalid input")
  @Response<ErrorResponse>(401, "Unauthorized")
  public async updateProfile(
    @Body() requestBody: { firstName: string; lastName: string },
    @Request() request: any,
  ): Promise<{ message: string } | ErrorResponse> {
    try {
      if (!requestBody.firstName && !requestBody.lastName) {
        this.setStatus(400);
        return {
          error: "Missing Fields",
          message: "At least one field (firstName or lastName) is required",
        };
      }

      // Get user from authorization header or session
      const authHeader = request.headers.authorization;
      if (!authHeader || !authHeader.startsWith("Bearer ")) {
        this.setStatus(401);
        return {
          error: "Unauthorized",
          message: "Authentication token required",
        };
      }

      const token = authHeader.substring(7);

      // Verify token and get user
      if (!this.supabaseService.client) {
        this.setStatus(500);
        return createErrorResponse("Authentication service not available", 500);
      }

      const {
        data: { user },
        error: authError,
      } = await this.supabaseService.client.auth.getUser(token);

      if (authError || !user) {
        this.setStatus(401);
        return {
          error: "Unauthorized",
          message: "Invalid authentication token",
        };
      }

      // Update user profile in Supabase
      const updateData: any = {};
      if (requestBody.firstName) updateData.first_name = requestBody.firstName;
      if (requestBody.lastName) updateData.last_name = requestBody.lastName;

      // For now, just simulate success since we can't easily update user metadata
      // without service account key
      console.log("Profile update request:", requestBody);
      console.log("Would update user metadata:", updateData);

      console.log("Profile updated successfully for user:", user.email);

      this.setStatus(200);
      return {
        message: "Profile updated successfully",
      };
    } catch (error) {
      console.error("Profile update error:", error);
      this.setStatus(500);
      return {
        error: "Internal Server Error",
        message: "Failed to update profile",
      };
    }
  }
}